+++ pipelines/toast_planck_ring_numbers.py
@@ -26,18 +26,18 @@ import pickle
 
 parser = argparse.ArgumentParser( description="Build toast/DPC ring number dictionaries.", fromfile_prefix_chars="@" )
 parser.add_argument( "--rimo", required=True, help="RIMO file" )
-parser.add_argument( "--freq", required=True, type=np.int, help="Frequency" )
+parser.add_argument( "--freq", required=True, type=int, help="Frequency" )
 parser.add_argument( "--dets", required=False, default=None, help="Detector list (comma separated)" )
 parser.add_argument( "--effdir", required=True, help="Input Exchange Format File directory" )
-parser.add_argument( "--obtmask", required=False, default=1, type=np.int, help="OBT flag mask" )
-parser.add_argument( "--flagmask", required=False, default=1, type=np.int, help="Quality flag mask" )
+parser.add_argument( "--obtmask", required=False, default=1, type=int, help="OBT flag mask" )
+parser.add_argument( "--flagmask", required=False, default=1, type=int, help="Quality flag mask" )
 parser.add_argument( "--ringdb", required=True, help="Ring DB file" )
-parser.add_argument( "--odfirst", required=False, default=None, type=np.int, help="First OD to use" )
-parser.add_argument( "--odlast", required=False, default=None, type=np.int, help="Last OD to use" )
-parser.add_argument( "--ringfirst", required=False, default=None, type=np.int, help="First ring to use" )
-parser.add_argument( "--ringlast", required=False, default=None, type=np.int, help="Last ring to use" )
-parser.add_argument( "--obtfirst", required=False, default=None, type=np.float, help="First OBT to use" )
-parser.add_argument( "--obtlast", required=False, default=None, type=np.float, help="Last OBT to use" )
+parser.add_argument( "--odfirst", required=False, default=None, type=int, help="First OD to use" )
+parser.add_argument( "--odlast", required=False, default=None, type=int, help="Last OD to use" )
+parser.add_argument( "--ringfirst", required=False, default=None, type=int, help="First ring to use" )
+parser.add_argument( "--ringlast", required=False, default=None, type=int, help="Last ring to use" )
+parser.add_argument( "--obtfirst", required=False, default=None, type=float, help="First OBT to use" )
+parser.add_argument( "--obtlast", required=False, default=None, type=float, help="Last OBT to use" )
 
 args = parser.parse_args()
 
@@ -148,7 +148,7 @@ for iring, (ring_start, ring_time, ring_len) in enumerate(
             raise Exception("This query {} corresponds to these PIDs: {}".format(cmd, pids))
         for pid in pids: pass
 
-        esa_id = np.int( pid )
+        esa_id = int( pid )
 
         cmd = "select LFI_ID, HFI_ID from rings where ESA_ID == {}".format(esa_id)
